Removed unnecessary casts.
git-svn-id: https://svn.apache.org/repos/asf/cayenne/main/trunk@762198 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/cayenne-jdk1.5-unpublished/src/main/java/org/apache/cayenne/wocompat/EOQuery.java
@@ -339,7 +339,7 @@ public class EOQuery extends SelectQuery {
          * @return int Expression type
          */
         static int expressionTypeForSelector(String selector) {
-            Integer expType = (Integer) selectorToExpressionBridge().get(selector);
+            Integer expType = selectorToExpressionBridge().get(selector);
             return (expType != null ? expType.intValue() : -1);
         }
